pam_sm_authenticate should cope with NULL token
In pam_sm_authenticate(), pam_get_item() can provide a token value of
NULL which will cause sscanf() to crash.  Check for this and return an
error code.

Note that a NULL token is not the same as a pointer to the empty string,
"" which seems to be the case for a blank password, so something
different is happening here.