Get authentication token if none is cached
Authentication succeeds (at least for me) with the PAM config in
unix-or-token-auth, which begins:

  auth	[success=2 default=ignore]	pam_unix.so nullok_secure
  auth	[success=1 default=ignore]	pam_transient_token.so

But if you reverse the order as below, or if you only specify
pam_transient_token.so, then the pam_get_item() gives a NULL auth token.

  auth	[success=2 default=ignore]	pam_transient_token.so
  auth	[success=1 default=ignore]	pam_unix.so nullok_secure

This is because pam_get_item() with PAM_AUTHTOK only looks at the cached
authentication token, i.e. the one set by a previous module, and it
doesn't actually obtain the token.  That has to be done by conversation;
I haven't yet implemented this for the general portable case but for
Linux it is simple to use the pam_get_authtok() helper function.

(Note that if you specify pam_transient_token.so first and pam_unix.so
second then you will get two prompts for the password unless you give
pam_unix.so the option try_first_pass.  However, pam_transient_token.so
always uses the cached token.)