Change auth string length to 16 bytes
An earlier commit changed the protocol from challenge-response to a
simple auth string, because the response is unnecessary.  However, the
PAM module did check the response and by removing it we effectively
halved the size of the random string.

This commit increases the auth string to 16 bytes to match the original
size of the random string.